pyproject.toml
[project]
name = "ai-jobverlust"
version = "0.1.0"
description = "Deutscher Arbeitsmarkt Visualizer - KI-Exposition"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.2.2",
]
